refactor FE helper files
